change EH related stuff (other than EH_LABEL) to use MCSymbol instead of label ID's.  This cleans up and regularizes a bunch of code and makes way for future progress.
Unfortunately, this pointed out to me that JITDwarfEmitter.cpp
is largely copy and paste from DwarfException/MachineModuleInfo
and other places.  This is very sad and disturbing. :(

One major change here is that TidyLandingPads moved from being
called in DwarfException::BeginFunction to being called in
DwarfException::EndFunction.  There should not be any
functionality change from doing this, but I'm not an EH expert.